+++ Content.Server/Chat/Managers/ChatManager.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using Content.Server.Administration.Logs;
 using Content.Server.Administration.Managers;
-using Content.Server.Corvax.Donations;
+using Content.Server.Corvax.Sponsors;
 using Content.Server.MoMMI;
 using Content.Server.Preferences.Managers;
 using Content.Server.Station.Systems;
@@ -38,7 +38,7 @@ namespace Content.Server.Chat.Managers
         [Dependency] private readonly IServerPreferencesManager _preferencesManager = default!;
         [Dependency] private readonly IConfigurationManager _configurationManager = default!;
         [Dependency] private readonly IEntityManager _entityManager = default!;
-        [Dependency] private readonly IDonationManager _donationManager = default!;
+        [Dependency] private readonly ISponsorsManager _sponsorsManager = default!;
 
         private StationSystem _stationSystem = default!;
 
@@ -181,10 +181,10 @@ namespace Content.Server.Chat.Managers
                 messageWrap = Loc.GetString("chat-manager-send-ooc-patron-wrap-message", ("patronColor", patronColor),("playerName", player.Name));
             }
 
-            var donatorColor = await _donationManager.GetDonatorOOCColor(player.UserId);
-            if (donatorColor != null)
+            var sponsorColor = await _sponsorsManager.GetSponsorOOCColor(player.UserId);
+            if (sponsorColor != null)
             {
-                messageWrap = Loc.GetString("chat-manager-send-ooc-patron-wrap-message", ("patronColor", donatorColor),("playerName", player.Name));
+                messageWrap = Loc.GetString("chat-manager-send-ooc-patron-wrap-message", ("patronColor", sponsorColor),("playerName", player.Name));
             }
 
             //TODO: player.Name color, this will need to change the structure of the MsgChatMessage
