+++ Content.Server/IoC/ServerContentIoC.cs
@@ -5,7 +5,7 @@ using Content.Server.Administration.Notes;
 using Content.Server.Afk;
 using Content.Server.Chat.Managers;
 using Content.Server.Connection;
-using Content.Server.Corvax.Donations;
+using Content.Server.Corvax.Sponsors;
 using Content.Server.Database;
 using Content.Server.EUI;
 using Content.Server.GhostKick;
@@ -58,7 +58,7 @@ namespace Content.Server.IoC
             IoCManager.Register<IAdminLogManager, AdminLogManager>();
             IoCManager.Register<PlayTimeTrackingManager>();
             IoCManager.Register<UserDbDataManager>();
-            IoCManager.Register<IDonationManager, DonationManager>();
+            IoCManager.Register<ISponsorsManager, SponsorsManager>();
         }
     }
 }
